requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
python-dotenv==1.0.1
pandas==2.2.3
scikit-learn==1.6.1
joblib==1.4.2
requests==2.32.3
gunicorn==23.0.0
